% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{episodeinfo}
\alias{episodeinfo}
\title{episodeinfo}
\format{
This data frame contains the following columns:
\describe{
  \item{\code{szn}}{Name of season}
  \item{\code{sznnumber}}{Season number}
  \item{\code{series}}{Top Chef US (listed as US); Top Chef US Masters
                       (listed as US Masters); Top Chef Canada (listed as
                       Canada)}
  \item{\code{overall.episode.number}}{Running number of episode within
                                       the series}
  \item{\code{episode}}{Episode number}
  \item{\code{episode_name}}{Name of episode}
  \item{\code{air_date}}{Date the episode originally aired}
  \item{\code{#.of.competitors}}{Number of Chefs still in the competition}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Top_Chef}
}
\usage{
data(episodeinfo)
}
\description{
A dataset containing information about each episode
}
\examples{
library(tidyverse)
episodeinfo \%>\% filter(szn=="World All Stars")
}
\keyword{datasets}
