% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{judges}
\alias{judges}
\title{judges}
\format{
This data frame contains the following columns:
\describe{
  \item{\code{szn}}{Name of season}
  \item{\code{sznnumber}}{Season number}
  \item{\code{series}}{Top Chef US (listed as US); Top Chef US Masters
           (listed as US Masters); Top Chef Canada (listed as Canada)}
  \item{\code{episode}}{Episode number}
  \item{\code{challenge_type}}{Challenge type: qualifying challenge,
             elimination, quickfire, sudden death quickfire, quickfire
             elimination, battle of the sous chefs}
  \item{\code{outcome_type}}{Is the challenge run as a team or as an
  individual?}
  \item{\code{guestjudge}}{Name of guest judge}
  \item{\code{competed_on_TC}}{Will have a value of Yes if they competed
  on a season of Top Chef}
  \item{\code{other_shows}}{Information about other shows that this
  individual has appeared on}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Top_Chef}
}
\usage{
data(judges)
}
\description{
A dataset containing information about who were the guest judges for
each challenge
}
\examples{
library(tidyverse)
judges \%>\%
  filter(guestjudge == "Eric Ripert") \%>\%
  group_by(challenge_type) \%>\%
  summarise(n=n())
}
\keyword{datasets}
