% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexCalculation_BasedOnNumOfChefs.R
\name{ScoresThroughAGivenNumberOfChefs}
\alias{ScoresThroughAGivenNumberOfChefs}
\title{Calculate One Season's Chefs' Weighted Scores Taking into Account
Number of Chefs Still in the Competition}
\usage{
ScoresThroughAGivenNumberOfChefs(
  seriesname,
  seasonnumberofchoice,
  numberofchefsofinterest,
  scoringsystem
)
}
\arguments{
\item{seriesname}{Values can be: US, US Masters, Canada}

\item{seasonnumberofchoice}{Integer of the season number within the series}

\item{numberofchefsofinterest}{The number of chefs you want to still be in
thecompetition. The score will include all challenges up to and including
that number of chefs}

\item{scoringsystem}{Values can be original or modified. Modified gives
points for just being in the competition; original is only for how they do in
the challenges.}
}
\value{
Tibble of index score for each contestant in that season and their
         placement
}
\description{
Calculates the Index score for each person within a season of
             Top Chef.
}
\details{
The result of elimination challenges and quickfire challenges are
         weighted.
         Scoring: Elimination win = +7 points. Elimination high = +3 points.
                  Elimination low = -3 points. Eliminated = -7 points.
                  Quickfire win = +4 points. Quickfire high = +2 points.
                  Quickfire low = -2 points.
         Combines Sudden Death Quickfires with Eliminations
         Excludes Qualifying Challenges
         Holding constant the number of elimination challenges and quickfire
         challenges will allow comparison across seasons if you want
}
