% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topChef-package.R
\docType{package}
\name{topChef-package}
\alias{topChef}
\alias{topChef-package}
\title{topChef: Top Chef Data}
\description{
Several datasets which describe the chef contestants in Top Chef, the challenges that they compete in, and the results of those challenges. This data is useful for practicing data wrangling, graphing, and analyzing how each season of Top Chef played out.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/celevitz/topChef}
  \item Report bugs at \url{https://github.com/celevitz/topChef/issues}
}

}
\author{
\strong{Maintainer}: Levitz Carly E \email{celevitz@gmail.com} [copyright holder]

}
\keyword{internal}
